backend/requirements.txt
Django>=4.0
djangorestframework>=3.14.0
tensorflow>=2.10.0
numpy
pandas
scikit-learn
joblib
gunicorn
